feat: add type overriding with type-fest for views
